+++ irs-policy-store/src/test/java/org/eclipse/tractusx/irs/policystore/services/PolicyStoreServiceTest.java
@@ -37,14 +37,15 @@ import java.time.OffsetDateTime;
 import java.util.Collections;
 import java.util.List;
 
-import org.eclipse.tractusx.irs.policystore.exceptions.PolicyStoreException;
 import org.eclipse.tractusx.irs.edc.client.policy.Constraint;
 import org.eclipse.tractusx.irs.edc.client.policy.Constraints;
-import org.eclipse.tractusx.irs.policystore.models.CreatePolicyRequest;
 import org.eclipse.tractusx.irs.edc.client.policy.OperatorType;
 import org.eclipse.tractusx.irs.edc.client.policy.Permission;
 import org.eclipse.tractusx.irs.edc.client.policy.Policy;
 import org.eclipse.tractusx.irs.edc.client.policy.PolicyType;
+import org.eclipse.tractusx.irs.policystore.config.DefaultAcceptedPoliciesConfig;
+import org.eclipse.tractusx.irs.policystore.exceptions.PolicyStoreException;
+import org.eclipse.tractusx.irs.policystore.models.CreatePolicyRequest;
 import org.eclipse.tractusx.irs.policystore.models.UpdatePolicyRequest;
 import org.eclipse.tractusx.irs.policystore.persistence.PolicyPersistence;
 import org.junit.jupiter.api.BeforeEach;
@@ -61,19 +62,19 @@ class PolicyStoreServiceTest {
 
     private static final String BPN = "testBpn";
     private static final String EXAMPLE_ALLOWED_NAME = "ID 3.0 Trace";
+    private static final String EXAMPLE_ACCEPTED_LEFT_OPERAND = "PURPOSE";
+    private final Clock clock = Clock.systemUTC();
+    @Captor
+    ArgumentCaptor<Policy> policyCaptor;
     private PolicyStoreService testee;
-
     @Mock
     private PolicyPersistence persistence;
 
-    @Captor
-    ArgumentCaptor<Policy> policyCaptor;
-
-    private final Clock clock = Clock.systemUTC();
-
     @BeforeEach
     void setUp() {
-        testee = new PolicyStoreService(BPN, List.of(EXAMPLE_ALLOWED_NAME), persistence, clock);
+        final DefaultAcceptedPoliciesConfig defaultAcceptedPoliciesConfig = new DefaultAcceptedPoliciesConfig();
+        defaultAcceptedPoliciesConfig.setAcceptedPolicies(List.of());
+        testee = new PolicyStoreService(BPN, defaultAcceptedPoliciesConfig, persistence, clock);
     }
 
     @Test
@@ -91,7 +92,8 @@ class PolicyStoreServiceTest {
     @Test
     void registerPolicyWithPermission() {
         // arrange
-        final var req = new CreatePolicyRequest("testId", OffsetDateTime.now(clock).plusMinutes(1), createPermissions());
+        final var req = new CreatePolicyRequest("testId", OffsetDateTime.now(clock).plusMinutes(1),
+                createPermissions());
 
         // act
         testee.registerPolicy(req);
@@ -131,14 +133,27 @@ class PolicyStoreServiceTest {
     }
 
     @Test
-    void getStoredPoliciesWhenEmpty() {
+    void getDefaultStoredPoliciesWhenEmpty() {
+        // arrange
+        final DefaultAcceptedPoliciesConfig.AcceptedPolicy acceptedPolicy1 = new DefaultAcceptedPoliciesConfig.AcceptedPolicy(
+                EXAMPLE_ACCEPTED_LEFT_OPERAND, "eq", EXAMPLE_ALLOWED_NAME);
+        final DefaultAcceptedPoliciesConfig.AcceptedPolicy acceptedPolicy2 = new DefaultAcceptedPoliciesConfig.AcceptedPolicy(
+                EXAMPLE_ACCEPTED_LEFT_OPERAND, "eq", EXAMPLE_ALLOWED_NAME);
+        final DefaultAcceptedPoliciesConfig defaultAcceptedPoliciesConfig = new DefaultAcceptedPoliciesConfig();
+        defaultAcceptedPoliciesConfig.setAcceptedPolicies(List.of(acceptedPolicy1, acceptedPolicy2));
+        testee = new PolicyStoreService(BPN, defaultAcceptedPoliciesConfig, persistence, clock);
+
         // act
         final var defaultPolicies = testee.getStoredPolicies();
 
         // assert
         assertThat(defaultPolicies).hasSize(1);
-        List<Permission> permissionList = defaultPolicies.get(0).getPermissions();
+        final List<Permission> permissionList = defaultPolicies.get(0).getPermissions();
         assertThat(permissionList).hasSize(1);
+        final List<Constraints> constraints = permissionList.get(0).getConstraints();
+        assertThat(constraints).hasSize(1);
+        assertThat(constraints.get(0).getOr()).hasSize(2);
+        assertThat(constraints.get(0).getAnd()).hasSize(2);
     }
 
     private Policy createPolicy(final String policyId) {
@@ -146,22 +161,16 @@ class PolicyStoreServiceTest {
     }
 
     private List<Permission> createPermissions() {
-        return List.of(
-                new Permission(PolicyType.USE, List.of(createConstraints())),
-                new Permission(PolicyType.ACCESS, List.of(createConstraints()))
-        );
+        return List.of(new Permission(PolicyType.USE, List.of(createConstraints())),
+                new Permission(PolicyType.ACCESS, List.of(createConstraints())));
     }
 
     private Constraints createConstraints() {
-        return new Constraints(
-                Collections.emptyList(),
-                List.of(
-                        new Constraint("Membership", OperatorType.EQ, List.of("active")),
+        return new Constraints(Collections.emptyList(),
+                List.of(new Constraint("Membership", OperatorType.EQ, List.of("active")),
                         new Constraint("FrameworkAgreement.traceability", OperatorType.EQ, List.of("active")),
-                        new Constraint("PURPOSE", OperatorType.EQ, List.of("ID 3.1 Trace")))
-        );
+                        new Constraint(EXAMPLE_ACCEPTED_LEFT_OPERAND, OperatorType.EQ, List.of("ID 3.1 Trace"))));
     }
-
 
     @Test
     void deletePolicy() {